refactor(sets): move DisableFormValueModelBindingAttribute to Shared Utilities
It's generic ASP.NET Core MVC infrastructure (a resource filter that
disables default form/file value providers for streaming multipart
uploads) with no Sets-specific logic, just like the neighboring
MultipartRequestHelper it's used alongside. Keep both together in
Utilities/ instead of putting the attribute under Sets/Attributes.
